Replacing still more uses of unsigned, 2.

+++ FindPanel/OgreTextFindComponents/NSOutlineView/OgreOutlineCellAdapter.m
@@ -101,7 +101,7 @@
 #endif
 }
 
-- (void)beginRegisteringUndoWithCapacity:(unsigned)aCapacity
+- (void)beginRegisteringUndoWithCapacity:(NSUInteger)aCapacity
 {
 }
